# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ImageUpload {
  imageToken: string;
}

export type ReconstructionRequestMode = typeof ReconstructionRequestMode[keyof typeof ReconstructionRequestMode];


export const ReconstructionRequestMode = {
  blueprint: 'blueprint',
  sketch: 'sketch',
  character: 'character',
  animal: 'animal',
  scan: 'scan',
} as const;

export interface ReconstructionRequest {
  mode: ReconstructionRequestMode;
  /**
     * @minItems 1
     * @maxItems 4
     */
  imageTokens: string[];
}

export type ReconstructionTaskStatus = typeof ReconstructionTaskStatus[keyof typeof ReconstructionTaskStatus];


export const ReconstructionTaskStatus = {
  queued: 'queued',
  running: 'running',
  success: 'success',
  failed: 'failed',
  banned: 'banned',
  expired: 'expired',
  cancelled: 'cancelled',
} as const;

export interface ReconstructionTask {
  taskId: string;
  status: ReconstructionTaskStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  modelPath: string | null;
  /** @nullable */
  message: string | null;
}

export type ModelExportRequestFormat = typeof ModelExportRequestFormat[keyof typeof ModelExportRequestFormat];


export const ModelExportRequestFormat = {
  GLTF: 'GLTF',
  USDZ: 'USDZ',
  FBX: 'FBX',
  OBJ: 'OBJ',
  STL: 'STL',
  '3MF': '3MF',
} as const;

export interface ModelExportRequest {
  format: ModelExportRequestFormat;
}