# lib/api-zod/src/generated/types/modelExportRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ModelExportRequestFormat } from './modelExportRequestFormat';

export interface ModelExportRequest {
  format: ModelExportRequestFormat;
}