# lib/api-zod/src/generated/types/reconstructionRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReconstructionRequestMode } from './reconstructionRequestMode';

export interface ReconstructionRequest {
  mode: ReconstructionRequestMode;
  /**
     * @minItems 1
     * @maxItems 4
     */
  imageTokens: string[];
}