# lib/api-zod/src/generated/types/reconstructionTask.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReconstructionTaskStatus } from './reconstructionTaskStatus';

export interface ReconstructionTask {
  taskId: string;
  status: ReconstructionTaskStatus;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  modelPath: string | null;
  /** @nullable */
  message: string | null;
}